Better handling of hangups when using mobile
The channel received from the cti server is not consistent from one call
to another.

When doing tests calling the same number from the same phone I got different
results that changed every one in a while.

Sometime we receive SIP/180000... and sometime Local/mobile number.
Hang-up doesnt work when the channel is Local/... so we hang-up the peer's
channel.