Added voice mail support
The voice mail button is now enabled. The button is bright when there's a
mwi warning. The button can be used to dial into the voice mail when _NOT_
using mobile. When using a mobile number the button will reflect the status
of the voice mail but doesn't allow to retrieve messages yet.