Keep the default incoming call dialer on mobile
When calling an external number from mobile number the CTI server
doesn't send channel ids. Without the channel ids, we cannot transfer
or hang-up a call.

The default behavior is to hide the android dialer to allow transfers
but since we can't hang-up and transfer the android dialer will stay
on the foreground while the call is taking place.